Add files via upload
We invented a unique way to edit code as it took too long to get a mutually editable code space up for the two of us. We had a word doc that we would change and put notes on, and then run the code in separate jupyter notebooks. The word doc is the code, and the pdfs are the outputs of the code. Let us know if you have any questions.